requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
aiohttp==3.9.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.0.0
python-multipart==0.0.6
google-cloud-aiplatform==1.36.4
requests==2.31.0
pytest==7.4.3
Flask==3.0.0
google-auth==2.23.4